Move to the right place this time: cmd/atenet/internal
This follows the convention documented here:

AGENTS.md

```
**Where to put new Go code — quick rules:**

| Only used by one binary | `cmd/<binary>/internal/<pkg>` |
```

See the discussion here: https://github.com/agent-substrate/substrate/issues/49
