+++ cmd/atenet/internal/dns.go
@@ -29,7 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 
-	"github.com/agent-substrate/substrate/internal/dns"
+	"github.com/agent-substrate/substrate/cmd/atenet/internal/dns"
 )
 
 type DnsConfig struct {
