+++ cmd/atenet/internal/dns/corefile.go
