+++ cmd/atenet/internal/router/dashboard.html
