+++ cmd/atenet/internal/router/status.go
